requirements.txt
aiofiles==22.1.0
aiogram==3.0.0b5
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.18.0
alembic==1.10.3
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
databases==0.7.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.86.0
frozenlist==1.3.3
googletrans==3.1.0a0
greenlet==2.0.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httptools==0.5.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.2.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
magic-filter==1.0.9
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.4
orjson==3.8.9
passlib==1.7.4
psycopg2==2.9.6
pyasn1==0.4.8
pydantic==1.10.7
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.47
SQLAlchemy-Utils==0.40.0
starlette==0.20.4
tqdm==4.65.0
transliterate==1.10.2
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uuid==1.30
uvicorn==0.18.3
watchfiles==0.19.0
webptools==0.0.9
websockets==11.0.1
yarl==1.8.2
zipp==3.15.0
